Fix handling of touch icons when it is disabled
Store the icon type in a bitfield because the same icon URL might be
used for various types on same page. This way webengine won't ignore to
download a default icon what is also set as a touch icon when touch
icons are disabled.

Moreover, do not store the icon types from the previous page because
a subsequent page might use the same icon URL but with different type.
With this change the type of the cached icons are updated after each
page load.

Fixes: QTBUG-70081
Change-Id: I8031a740b07b0c6a8e5759a994f386b13ce87be2
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>