Fix crash in tst_Spellchecking::spellcheck
The test tst_Spellchecking::spellcheck crashes when run with

    QT_LOGGING_RULES=qt.accessibility.cache.debug=true

which triggers debug output from

    QAccessible::registerAccessibleInterface and
    QAccessible::deleteAccessibleInterface

These functions are called during the construction/destruction of
BrowserAccessibilityQt. The problem is that BrowserAccessibilityQt is not
necessarily in a valid state at these times: BrowserAccessibilityQt relies on
it's manager to be *fully* initialized yet BrowserAccessibilityQt may be
constructed/destructed *during* the managers construction/destruction.

The solution is to override QAccessibleInterface::isValid to return true only if
the manager is in a valid state too. This disables the debug output and avoids
the crash.

Fixes: QTBUG-72878
Change-Id: Icffe4e503145133c24535fb094bbcb606568e57c
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>