Fix default profile for single-process mode
In single process mode there is only one profile,
and this profile is used for browser context.
Fix the case where there is no 'default' profile
created but only one user profile. In this case
make the user profile 'default' so it is deleted
in right moment in single process mode.

Change-Id: I8ba3c2da2d93d53fa569c7971410a4a74753377e
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>