Persistent storage of preferences between app runs
We can now persist the user preferences to disk at application exit so
they are correctly remembered for the next run. The preferences are
stored in dataPath()/user_prefs.json. This is managed via the
PrefServiceAdapter class.

It's not possible to test this in a correct pattern as we require the
app to exit properly to be able to guarantee that the preferences are
written to disk. The code added in ~PrefServiceAdapter() does not
provide any guarantees in the next profile construction that the
previous preferences have been written to disk.

Fixes: QTBUG-75740
Change-Id: I96bb3037707d90b05b3dfab4f083702351923aee
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>