Fix QWebEnginePage visibility state
Calling

  view.show();
  view.load(...);
  view.hide();

results in a WebContents in visible state because the first call to show() is
delayed via a timer and never canceled so it takes effect after the hide().

Also, the page's visibility state is not updated when changing views, so that a
visible QWebEngineView can be set to display a hidden QWebEnginePage.

Fixes: QTBUG-77209
Change-Id: I0243c659f4039a35c36e04ea55c5972666170fd0
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>