Add webengine-core feature
Create/split main configure for submodules. Now configure system knows when webenginecore
module is not going to be built and the user gets the feedback after the configure step
with the message:

The following modules are not being compiled in this configuration:
    webenginecore
    webengine
    webenginewidgets

If a module is not built also features are not populated, therefore some tests
have to be moved to main configuration. This improves error reporting and prepares
landing for QtPdf modules.

Remove configure.prf and fix issues after config split.

Add makefile call to report errors. Now calling make also reports
issues.

Task-number: QTBUG-75840
Task-number: QTBUG-76606
Change-Id: I76944df4c5db6f4954c464e3741a8054cb10b40e
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>