WebContentsAdapter: don't use toStrongRef().data()
It's an anti-pattern.

Even though we don't expect other threads to destroy the last QSP
while we use the pointer obtained from QWeakPointer, play it safe
and use QWeakPointer idiomatically: as a non-owning reference, to
be converted to an owning one for the duration of our use of the
payload object.

Amends 7cade5688f6a0bde067dc496cedcb8de14662116.

Change-Id: Ib1c136569c27f413a4a78bfc35784032c53522a9
Reviewed-by: Michael Brüning <michael.bruning@qt.io>