Add a QImageIOHandler to render PDF with QtQuick Image
Now you can use Image in QML to display the first page of a PDF file
by default. (The implementation is very similar to the handler
plugin in the QtSVG module.) To display other pages, you can set
the new currentFrame property that is being added to Image in 5.14.

QPdfIOHandler uses its own instance of QPdfDocument to do the rendering.
An instance will be created each time the image needs to read a frame
from the file (on creation; each time the currentFrame property
is changed; each time sourceSize is changed to require a different
resolution rendering). This approach is not as efficient as it would
be to share the QPdfDocument instance among multiple Images that are
rendering different pages at the same time, as well as sharing with the
QML Document declaration that we plan to add. However, if you set
asynchronous: true, each Image will render its page from the PDF in
its own thread. PDFium is not thread-safe, so sharing the QPdfDocument
among multiple threads is not safe. It may be possible to make it safe
by having a dedicated PDF-rendering thread, but that is not yet done
in this patch.

Change-Id: I7a1f8cd7bd5b8f93d45fa9350752b2d9356b04fe
Reviewed-by: Michal Klocek <michal.klocek@qt.io>