PDF viewer and manual test: support password-protected documents
The multipage example was OK already; this adds the same password-prompt
dialog to the single-page-at-a-time pdfviewer example and to the
withdoc.qml manual test. Also, the dialogs seem to size themselves
better if we explicitly set the contentItem properties. And an
error dialog should not have an OK button but a Close button:
the error is not OK, but the example does not provide a way to
try to fix it either.

Fixes: QTBUG-83988
Fixes: QTBUG-96574
Change-Id: I9d8c3cbe1f955cea4319982869ec9180d7c8666c
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>