Clean up platform tests
Make it a private feature so test is not run twice.
The better solution would be to split the platform
test into separate tests in configure, but it looks like
a lot of boring typing and would result in bunch on private
features. Instead run test only on configure step
and in case of fail just export the error, so it is
reported during configure and make calls.

Change-Id: Ice1188631a53fe739569de3e303533b18342941a
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>