Fix life cycle issue of QWebEngineUrlRequestInterceptor
Since 5e92adf intercept() is called on ui thread, however
this patch does not consider deletion of interceptor while
io thread still process the request and for examples
checks interceptor's deprecated property. Fix it.

Note this only fixes issue for interceptors which are
not deprecated therefore run on ui thread.

Change-Id: I5d5909065563e57a0cacb81fd04271b3f88596de
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>