Don't modify the allowed actions as they are correct already
By not modifying the allowed actions instead of just setting it to be the
MoveAction will enable dragging from a webpage to another webpage to
work correctly. Otherwise it will potentially reject the drag because
it is seen only as a move, whereas it wouldn't be possible to really
move from one webpage to another as the original source cannot be
deleted.

Change-Id: I34105d10e7d1dc831016c33c9c6cfc544c4e084b
Reviewed-by: Jüri Valdmann <juri.valdmann@qt.io>