Fix getDisplayMedia crash
The MEDIA_DISPLAY_VIDEO_CAPTURE stream type is handled incorrectly by
MediaCaptureDevicesDispatcher causing a crash when an unexpected type of media
device is returned to Chromium.

This patch only fixes the crash, screen sharing is nonetheless not properly
supported by WebEngine due to limitations of the public API which does not allow
selecting between screens, not to mention windows or tabs.

On Linux WebRTC's ScreenCapturer is not even built since it depends on use_x11
being set in GN.

Fixes: QTBUG-78016
Change-Id: I7fa49febaba1be94bdb6c31265dfc24ee809d635
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>