Redo scale handling of pdf when printing
Since 7f89bad our printer handling produces low quality
images, the reason for that is the we first render
with pdfium to bitmap which is fixed to twice size of pdf
document in points, then it is downscaled/upscaled by
qpainter. This issue is even more visible when print preview
is requested on hdpi screen.

Simplify the code and do the "scaling" already when rendering
bitmap with pdfium.

Task-number: QTBUG-75092
Fixes: QTBUG-78161
Change-Id: I616ad1f371e7d3457a04b0a254603ed235c387bc
Reviewed-by: Michael Brüning <michael.bruning@qt.io>