Store favicon URL when serializing QWebEngineHistory
This is to make QWebEngineHistoryItem::iconUrl of the deserialized
QWebEngineHistory object return the same URL it had when serializing. Otherwise
it's impossible to have favicons for URLs from the navigation history of a
restored browser session until these URLs are visited again
(so that Chromium refetches their favicons). These icons are usually needed much
earlier - e.g. when showing a popup with the navigation history and having an
icon URL allows to load it from a disk cache before visiting the page.

Fixes: QTBUG-78998
Change-Id: Ief2d089d52f301826e5c131d401cafd08952a8b5
Reviewed-by: Michael Brüning <michael.bruning@qt.io>