Add entitlements to QtWebEngineProcess.app
Adds entitlements needed by the QtWebEngineProcess.app bundle
when running with the hardened runtime.

To use these entitlements when deploying an application with the
hardened runtime enabled or notarizing it.

To use it, deploy and sign the application using macdeployqt, then
re-sign it using the additional entitlements that are contained in
this patch. Coming version of macdeployqt may do this automatically
when the hardened runtime is enabled.

Task-number: QTBUG-77442
Change-Id: I7fde358725cf2d490f5e644934bf6466c2b4e1cb
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>