Stop using loadVisuallyCommitted in tst_qquickwebengineviewgraphics
The loadVisuallyCommitted signal is not emitted until the grabWindow call, but
the grabWindow call is not made until the loadVisuallyCommitted signal is
emitted. This could maybe work in Prolog but in C++ it's not a good idea.

Additionally, support not only images in RGB32 format, but also ARGB32, since
that's what we get with software compositing.

Fixes: QTBUG-58449
Task-number: QTBUG-79626
Change-Id: Iddae69764855febbc3a985ef7009227bc94634a5
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>