Fix renderProcessTerminated signal
With the adaptations for Chromium 76, RenderWidgetHostViewQt was changed to
become a RenderProcessHostObserver with the renderProcessTerminated signal being
emitted from the override of RenderProcessHostObserver::RenderProcessExited.

The problem with this can be seen by setting a breakpoint on the
RenderProcessGone override in RenderWidgetHostViewQt. We then get the trace:

  QtWebEngineCore::RenderWidgetHostViewQt::RenderProcessGone
  content::RenderWidgetHostImpl::RendererExited()
  content::RenderViewHostImpl::RenderProcessExited
  content::RenderProcessHostImpl::ProcessDied

ProcessDied iterates over all the observers and calls RenderProcessExited. Both
the RenderViewHostImpl and our RWHVQt are observers, but the RVHImpl comes
first. The RVHImpl then calls RendererExited, which calls our RenderProcessGone,
which does a 'delete this'. Now our RenderProcessExited override can never be
called because we have already deleted our observer.

Fix by moving the RenderProcessGone code to WebContentsDelegateQt and getting
the exit code from WebContents::GetCrashedErrorCode. Also add test.

Task-number: QTBUG-80085
Change-Id: I434744286df97a37b64722d7c15a1d4ee11c8af6
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>