Update find request id when a new search interrupts an ongoing search
If the new test became flaky it might happen because the first text
search finished before the second findText() call. This is very
unlikely, but in this case the test should be modified to not to check
if the first find failed. The point is to check we get the correct
amount of signals and the second search doesn't assert.

If the callbacks will be removed in Qt6, it should be re-considered to
remove the "unfinished find" workaround and trigger the first successful
findTextFinished() signal even if it happens in the middle of another
search.

Fixes: QTBUG-80086
Change-Id: I9c1ce20fc43fd81e8af784385a00ac2e7f7603b7
Reviewed-by: Kirill Burtsev <kirill.burtsev@qt.io>