Fix loading resources over qrc
The network service patch causes some test failures (with network
service *disabled*) because of the new method override IsHandledURL that was
added to ContentBrowserClientQt. The problem is that CanRequestURL in
ChildSecurityPolicyImpl forbids child processes from requesting URLs which
should be handled by the browser process, that is, URLs for which IsHandledURL
is true. Fix by explicitly granting permission to request custom schemes.

Change-Id: I970061de7ffb224de6121b85c1d5025a76f71e9a
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>