Fix build: remove dependency on Qt private header
The size of QJsonDocument's binary format is well known and we don't
need the header to know what it is. This fixes the build with 5.15,
where the contents of the previous QJsonPrivate namespace are now
in QBinaryJsonPrivate.

web_channel_ipc_transport_host.cpp:148:51: error: 'Header' is not a member of 'QJsonPrivate'

Change-Id: Id7decde0c426479bbf61fffd15dcc5c20a9eca2c
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>