QPdfDocument: check first page in load(), current page in render()
When a viewer application is opened with a book-sized PDF, there's no
need to check hundreds of pages before we can render the first page.
Using it in QPdfIOHandler is even worse because this check needs to
be repeated each time we advance from one page to the next.  But we
do need to ensure that the page we want to render is available.
Amends 945840bd067d9ca3179a667f48b451cc2087931b

Change-Id: Ib6576b1b91c63c2b57893d14b05632eff8cc4a15
Reviewed-by: Michal Klocek <michal.klocek@qt.io>