Add QQuickPdfPageImage and use in the PDF viewer components
Image works fine, except that if QPdfIOHandler is not given an
already-allocated QPdfDocument instance, it needs to construct its own.
So we now have QPdfFile: a subclass of QFile acting as a wrapper, to
carry the document instance (that the user has most likely declared)
down into QPdfIOHandler::load(QIODevice *). Thus, in nontrivial
PDF-viewing use cases, there is now usually only one multi-purpose
document instance. And this takes care of viewing password-protected
PDFs in the multipage example, because we already prompt for it
and set the document.password property.

In trivial use cases, it's OK to continue using Image, and QPdfIOHandler
will still construct its own QPdfDocument instance if the QIODevice
cannot be cast to a QPdfFile.

Task-number: QTBUG-77506
Task-number: QTBUG-83988
Task-number: QTBUG-96574
Change-Id: I3adfa54c30b0baa5dedebcf3bc759758f136b757
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>