Fix overriding shortcuts in password input fields on Windows
Windows IME does not support hidden text therefore IME input is disabled
on password fields. The shortcuts are supposed to be overridden in input
fields. Checking the keyboard focus on an input field is done by
verifying if the IME is enabled. This won't work with password fields on
platforms where hidden text is not supported, so also check if the
Qt::ImhHiddenText IME hint is set.

Fixes: QTBUG-81206
Change-Id: I81870beb556a9dda67295496dad8b672fbc5eba2
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>