QPdfDocument::pageSize(): add checkPageComplete()
checkPageComplete() calls FPDFAvail_IsPageAvail(), which does some
initialization that is necessary before PDFium is able to provide any
information about that page.  The static readImage() function in
qquickpixmapcache.cpp calls QImageReader::size() before rendering,
to determine the size of the image that will be read, and that
ends up calling QPdfDocument::pageSize(); so it's important to
"go to" that page before attempting to read its size.  In summary,
checkPageComplete() is the "go to page" function that must be called
before any other page-specific function.

Amends 9231d3444555945297857ee4aae05919083ea479

Change-Id: Idf5ba9d013b62395c284d055fbfe835fe949df79
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>