Fix PdfLinkModel's location y coordinate; add PdfMultiPageView tooltip
As usual, coordinates are in the first quadrant, and we need to
convert to 4th quadrant to get a y value that can be used to adjust
contentY of a ListView or TableView.

The tooltip when hovering over links provides a way to verify that
the link really jumps where it's intended to.

Change-Id: I9107639f15496a987c0fa7c3c2e2583c3839cc6b
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>