PdfScrollablePageView: retain position after pinch zoom
If you do the pinch on a trackpad, the mouse cursor doesn't move, and
the same visible part of the page will remain under the cursor after
the page is re-rendered. Likewise when doing the pinch on a touchscreen,
the centroid (midpoint between the two fingers) is held in place.
This is achieved by moving the scrollbars afterwards, subject to their
constraints.

Change-Id: I34caca4ebbb5c11007dd2462416a42c1a2d8e104
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>