PdfMultiPageView: retain position after pinch zoom
If you do the pinch on a trackpad, the mouse cursor doesn't move, and
the same visible part of the page will remain under the cursor after the
page is re-rendered.  Likewise when doing the pinch on a touchscreen,
the centroid (midpoint between the two fingers) is held in place,
subject to the constraints that Flickable.returnToBounds() imposes.
Similar to 29cb44ee471908ac7c4cac70e3defb8bd72a80cd.

Also corrected the horizontal scrolling range according to rotation;
added left and right margins so that the page edges are more visible
when the page is wider than the view, and thus the horizontal scrolling
range feels more accurate while the scrollbars are still shown.

Change-Id: Ia2a15eee5bcd5c9e7f025634f02a5a546e6aab68
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>