Suppress error message on ACCESSIBILITY_EVENTS permission type
The ACCESSIBILITY_EVENTS permission is used to enable AOM
(Accessibility Object Model) event listeners in blink. The current
implementation of the AOM is deprecated and it doesn't seem to be
supported in the foreseeable future by QtWebEngine.

Avoid the "Not implemented" error message in case of this permission
type because it is kept unsupported on purpose but blink still registers
permission status listener when AXObjectCache is created.

Change-Id: I4e9babb06015635e6c4c94c8fe433c714329692b
Reviewed-by: Kirill Burtsev <kirill.burtsev@qt.io>