PDF views: jump to precise locations
PdfScrollingPageView and PdfMultiPageView were already jumping to the
right page; but if you have zoomed in very far, you need it to jump
to the right part of the page too.

This affects how it jumps to search results, link locations, and the
forward/back behavior.  All of those should be more precise and
repeatable now.  But we depend on some new features that are added to
TableView for Qt 6; in lieu of those, we use TableViewExtra for now.

Fixes: QTBUG-83679
Change-Id: Ie974205562fe7dbf93bae274cef6fefa768aaefb
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>