Add PdfStyle; use Control palette colors
PdfStyle.qml can be overridden via file selectors, and contains extra
styling beyond what Control.palette provides.

Search results are highlighted with a lightened, translucent version of
the Controls style's accent color, and the current search result is
highlighted with a solid border with the accent color.

The text selection color comes from the Controls theme highlight color,
which tends to be the same as QPalette::Highlight.

The link underscore is also configurable in PdfStyle, but its color
comes from the Controls theme link color, which tends to be the same as
QPalette::link.

Task-number: QTBUG-82540
Change-Id: I7c7e80323cf5d7c94dde775a38cdec239a6351a9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>